refactor(toolbar): extract FocusCycleManager from toolbar-ui.js into its own module
